[standard][perf] Convert CJS requires in react-native-implementation to ES imports/exports
This is the first step towards converting RN to use ECMAScript modules (ESM) instead of CommonJS modules. Most of RN can be converted in a straightfoward manner, but the react-native-implementation entry point uses getters containing lazy `require` calls. This commit is one of two parts to move to JS modules while preserving the lazy require behavior. The second part is in metro-react-native-babel-preset.

This commit changes all of the `require` calls in react-native-implementation.js to `export from` statements. Some of the old `require` calls had deprecation warnings in their enclosing getters -- these warnings are now in the deprecated modules themselves. (AlertIOS is one exception since is is a deprecated alias for Alert. We could add an intermediate AlertIOS.js file that prints the warning but I opted to keep the getter for this one module since we are removing it anyway.)

The Metro side of this conversion enables the `lazy` option for `@babel/plugin-transform-modules-commonjs` when any of the RN modules are required. This means that lazy `require` calls in react-native-implementation are preserved out of the box. (And if you specify `lazyImportExportTransform: true`, all imports are lazy and there's no change in behavior.)

Test Plan:

Opened RNTester and verified it loaded without an error. Tapped on several of the test screens. Also got yellow box warnings for the deprecated modules, since RNTester loads them.

Locally changed RNTester to be a very simple app that just displays a string. Verified the synthetic default export still works with `require()` and that no deprecation warnings were displayed because only non-deprecated components were used and the imports are lazy:
```js
const React = require('react');
const ReactNative = require('react-native');
const {AppRegistry, StyleSheet, Text, View} = ReactNative;
```

Also double-checked that the Metro Babel preset preserves lazy `require` calls in react-native-implementation.js by looking at the code in the bundle:

```js
__d(function (global, _$$_REQUIRE, _$$_IMPORT_DEFAULT, _$$_IMPORT_ALL, module, exports, _dependencyMap) {
  'use strict';

  var _interopRequireDefault = _$$_REQUIRE(_dependencyMap[0], "@babel/runtime/helpers/interopRequireDefault");

  Object.defineProperty(exports, "__esModule", {
    value: true
  });
  Object.defineProperty(exports, "AccessibilityInfo", {
    enumerable: true,
    get: function get() {
      return _AccessibilityInfo().default;
    }
  });

  // ... (one defineProperty call with a deferred getter for each export)

  var _extends2 = _interopRequireDefault(_$$_REQUIRE(_dependencyMap[1], "@babel/runtime/helpers/extends"));

  function _AccessibilityInfo() {
    var data = _interopRequireDefault(_$$_REQUIRE(_dependencyMap[2], "AccessibilityInfo"));

    _AccessibilityInfo = function _AccessibilityInfo() {
      return data;
    };

    return data;
  }

  // ... (one lazy, memoized function for each import)

  (0, _extends2.default)(exports, 'AlertIOS', {
    get: function get() {
      var warnOnce = _$$_REQUIRE(_dependencyMap[94], "warnOnce");

      warnOnce('alert-ios', 'AlertIOS is deprecated. Use the `Alert` module directly instead.');
      return _$$_REQUIRE(_dependencyMap[46], "Alert");
    }
  });
},317,[1,13,318,319,357,367,451,452,454,413,432,455,457,458,459,461,462,468,470,511,513,355,515,493,420,437,517,519,521,417,523,524,528,368,530,537,373,482,380,381,442,341,538,414,539,542,264,382,543,307,312,314,553,554,555,324,323,409,277,477,556,390,423,424,316,249,557,526,268,322,558,559,560,561,321,183,562,563,315,280,500,564,491,250,273,187,229,299,344,565,327,370,566,483,313],"Libraries/react-native/react-native-implementation.js");
```